tests/wpt/web-platform-tests/presentation-api/controlling-ua/PresentationRequest_success.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Presentation API PresentationRequest for Controlling User Agent (Success)</title>
<link rel="author" title="Franck William Taffo" href="http://www.fokus.fraunhofer.de">
<link rel="help" href="http://w3c.github.io/presentation-api/#dfn-controlling-user-agent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
    test(function() {
        try {
            var request = new PresentationRequest('presentation.html');
            assert_true(request instanceof PresentationRequest);
        }
        catch (ex) {
            assert_unreached('PresentationRequest constructor threw an unexpected exception "' + ex.name + '"');
        }
    }, 'Call PresentationRequest constructor with a valid relative presentation URL. No Exception expected.');

    test(function() {
        try {
            var request = new PresentationRequest('http://example.org/');
            assert_true(request instanceof PresentationRequest);
        }
        catch (ex) {
            assert_unreached('PresentationRequest constructor threw an unexpected exception "' + ex.name + '"');
        }
    }, 'Call PresentationRequest constructor with a valid absolute presentation URL. No Exception expected.');

    test(function() {
        try {
            var request = new PresentationRequest([
                'presentation.html',
                'http://example.org/presentation/'
            ]);
            assert_true(request instanceof PresentationRequest);
        }
        catch (ex) {
            assert_unreached('PresentationRequest constructor threw an unexpected exception "' + ex.name + '"');
        }
    }, 'Call PresentationRequest constructor with a set of valid presentation URLs. No Exception expected.');

</script>
